requirements.txt
aiodns==3.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
argcomplete==3.6.2
arrow==1.3.0
attrs==25.1.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
binaryornot==0.4.4
bottle==0.13.4
briefcase==0.3.25
Brotli==1.1.0
build==1.3.0
cbor2==5.6.5
certifi==2025.7.9
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.3.0
cookiecutter==2.6.0
cryptography==44.0.0
cupshelpers==1.0
Cython==3.0.12
dasbus==1.7
dbus-python==1.3.2
distro==1.9.0
dnf==4.23.0
evdev==1.9.1
fastapi==0.117.1
filelock==3.15.4
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
i3ipc==2.2.1
idna==3.10
ifaddr==0.1.7
importlib_metadata==8.6.1
inputs==0.5
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
ldap3==2.9
legacy-cgi==2.6.2
libcomps==0.1.22
libdnf==0.74.0
libvirt-python==11.0.0
lutris==0.5.19
lxml==5.3.2
lz4==4.3.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mercurial==6.9.5
moddb==0.12.0
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
mutagen==1.47.0
netifaces==0.11.0
nftables==0.1
numexpr==2.10.2
numpy==2.2.6
nwg-panel==0.10.5
olefile==0.47
packaging==24.2
pefile==2024.8.26
perf==0.1
pillow==11.1.0
platformdirs==4.4.0
ply==3.11
progressbar2==4.5.0
prompt_toolkit==3.0.52
protobuf==3.19.6
proton-core==0.4.0
proton-keyring-linux==0.2.0
proton-vpn-api-core==0.42.5
proton-vpn-gtk-app==4.9.7
proton-vpn-network-manager==0.12.14
protontricks==1.12.1
proxy_tools==0.1.0
psutil==6.1.1
pyasn1==0.6.1
pycairo==1.25.1
pycares==4.10.0
pycparser==2.20
pycryptodomex==3.23.0
pycups==2.0.4
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyGObject==3.50.0
pyinotify==0.9.6
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyNaCl==1.5.0
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyOpenSSL==25.0.0
pyotp==2.9.0
pypresence==4.3.0
pyproj==3.7.0
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5_sip==12.16.1
PySocks==1.7.1
pystray==0.19.5
python-dateutil==2.9.0.post0
python-gnupg==0.5.4
python-linux-procfs==0.7.3
python-multipart==0.0.20
python-slugify==8.0.4
python-utils==3.9.0
python-xlib==0.33
pyudev==0.24.3
pywebview==6.0
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.16.7-build/kernel-6.16.7/linux-6.16.7-200.fc42.x86_64/tools/net/ynl
pyzstd==0.16.2
referencing==0.36.2
rencode==1.0.6
requests==2.32.4
rich==14.1.0
rpds-py==0.25.0
rpm==4.20.1
rpmautospec==0.8.1
rpmautospec-core==0.1.5
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
sentry-sdk==2.21.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setuptools==74.1.3
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.48.0
sympy==1.14.0
systemd-python==235
text-unidecode==1.3
tomli_w==1.2.0
truststore==0.10.4
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.12.2
umu-launcher==1.2.9
urllib3==2.3.0
uvicorn==0.36.0
vdf==3.4
wcwidth==0.2.13
websockets==14.1
wheel==0.45.1
xpra==5.0.10
xxhash==3.5.0
yarl==1.13.1
yt-dlp==2025.9.5
zeroconf==0.118.0
zipp==3.21.0
zstandard==0.23.0
